Changed iterator types to unsigned int

+++ 0x1A-sorting_algorithms/0-bubble_sort.c
@@ -10,7 +10,7 @@
 
 void bubble_sort(int *array, size_t size)
 {
-	size_t i, j;
+	unsigned int i, j;
 	unsigned int temp;
 
 	if (size < 2 || !array || !size)
